+++ pages/instructorQuestionSettings/instructorQuestionSettings.js
@@ -15,11 +15,11 @@ const {
   QuestionDeleteEditor,
   QuestionCopyEditor,
 } = require('../../lib/editors');
-const config = require('../../lib/config');
+const { config } = require('../../lib/config');
 const sql = sqldb.loadSqlEquiv(__filename);
 const { encodePath } = require('../../lib/uri-util');
 const { idsEqual } = require('../../lib/id');
-const csrf = require('../../lib/csrf');
+const { generateSignedToken } = require('@prairielearn/signed-token');
 
 router.post('/test', function (req, res, next) {
   // We use a separate `test/` POST route so that we can always use the
@@ -83,7 +83,7 @@ router.post('/test', function (req, res, next) {
 router.post('/', function (req, res, next) {
   if (req.body.__action === 'change_id') {
     debug(`Change qid from ${res.locals.question.qid} to ${req.body.id}`);
-    if (!req.body.id) return next(new Error(`Invalid QID (was falsey): ${req.body.id}`));
+    if (!req.body.id) return next(new Error(`Invalid QID (was falsy): ${req.body.id}`));
     if (!/^[-A-Za-z0-9_/]+$/.test(req.body.id)) {
       return next(
         new Error(
@@ -227,7 +227,7 @@ router.get('/', function (req, res, next) {
 
   // Generate a CSRF token for the test route. We can't use `res.locals.__csrf_token`
   // here because this form will actually post to a different route, not `req.originalUrl`.
-  const questionTestCsrfToken = csrf.generateToken(
+  const questionTestCsrfToken = generateSignedToken(
     { url: questionTestPath, authn_user_id: res.locals.authn_user.user_id },
     config.secretKey
   );