+++ server.js
@@ -31,9 +31,15 @@ const { filesize } = require('filesize');
 const url = require('url');
 const { createProxyMiddleware } = require('http-proxy-middleware');
 const compiledAssets = require('@prairielearn/compiled-assets');
+const {
+  SCHEMA_MIGRATIONS_PATH,
+  initBatchedMigrations,
+  startBatchedMigrations,
+  stopBatchedMigrations,
+} = require('@prairielearn/migrations');
 
 const { logger, addFileLogging } = require('@prairielearn/logger');
-const config = require('./lib/config');
+const { config, loadConfig, setLocalsFromConfig } = require('./lib/config');
 const load = require('./lib/load');
 const awsHelper = require('./lib/aws.js');
 const externalGrader = require('./lib/externalGrader');
@@ -46,7 +52,6 @@ const error = require('@prairielearn/error');
 const sprocs = require('./sprocs');
 const news_items = require('./news_items');
 const cron = require('./cron');
-const redis = require('./lib/redis');
 const socketServer = require('./lib/socket-server');
 const serverJobs = require('./lib/server-jobs');
 const freeformServer = require('./question-servers/freeform.js');
@@ -114,7 +119,7 @@ module.exports.initExpress = function () {
     next();
   });
   app.use(function (req, res, next) {
-    config.setLocals(res.locals);
+    setLocalsFromConfig(res.locals);
     next();
   });
 
@@ -1369,44 +1374,30 @@ module.exports.initExpress = function () {
   );
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/clientFilesQuestion',
-    [
-      require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./middlewares/studentAssessmentAccess'),
-      require('./pages/clientFilesQuestion/clientFilesQuestion'),
-    ]
+    require('./pages/clientFilesQuestion/clientFilesQuestion')
   );
 
   // generatedFiles
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/generatedFilesQuestion',
-    [
-      require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./middlewares/studentAssessmentAccess'),
-      require('./pages/generatedFilesQuestion/generatedFilesQuestion'),
-    ]
+    require('./pages/generatedFilesQuestion/generatedFilesQuestion')
   );
 
   // Submission files
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/submission/:submission_id/file',
-    [
-      require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./middlewares/studentAssessmentAccess'),
-      require('./pages/submissionFile/submissionFile'),
-    ]
+    require('./pages/submissionFile/submissionFile')
   );
 
   // legacy client file paths
-  app.use('/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/file', [
-    require('./middlewares/selectAndAuthzInstanceQuestion'),
-    require('./middlewares/studentAssessmentAccess'),
-    require('./pages/legacyQuestionFile/legacyQuestionFile'),
-  ]);
-  app.use('/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/text', [
-    require('./middlewares/selectAndAuthzInstanceQuestion'),
-    require('./middlewares/studentAssessmentAccess'),
-    require('./pages/legacyQuestionText/legacyQuestionText'),
-  ]);
+  app.use(
+    '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/file',
+    require('./pages/legacyQuestionFile/legacyQuestionFile')
+  );
+  app.use(
+    '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/text',
+    require('./pages/legacyQuestionText/legacyQuestionText')
+  );
 
   //////////////////////////////////////////////////////////////////////
   //////////////////////////////////////////////////////////////////////
@@ -1671,6 +1662,10 @@ module.exports.initExpress = function () {
     '/pl/administrator/courseRequests/',
     require('./pages/administratorCourseRequests/administratorCourseRequests')
   );
+  app.use(
+    '/pl/administrator/batchedMigrations',
+    require('./pages/administratorBatchedMigrations/administratorBatchedMigrations')
+  );
 
   //////////////////////////////////////////////////////////////////////
   //////////////////////////////////////////////////////////////////////
@@ -1857,11 +1852,9 @@ if (config.startServer) {
           configFilename = argv['config'];
         }
 
-        // Load config values from AWS as early as possible so we can use them
-        // to set values for e.g. the database connection
-        await config.loadConfigAsync(configFilename);
+        // Load config immediately so we can use it configure everything else.
+        await loadConfig(configFilename);
         await awsHelper.init();
-        await awsHelper.loadConfigSecrets();
 
         // This should be done as soon as we load our config so that we can
         // start exporting spans.
@@ -1987,16 +1980,34 @@ if (config.startServer) {
 
         // Our named locks code maintains a separate pool of database connections.
         // This ensures that we avoid deadlocks.
-        await namedLocks.init(pgConfig, idleErrorHandler);
+        await namedLocks.init(pgConfig, idleErrorHandler, {
+          renewIntervalMs: config.namedLocksRenewIntervalMs,
+        });
 
         logger.verbose('Successfully connected to database');
+      },
+      async () => {
+        // We need to do this before we run migrations, as some migrations will
+        // call `enqueueBatchedMigration` which requires this to be initialized.
+        const runner = initBatchedMigrations({
+          project: 'prairielearn',
+          directories: [path.join(__dirname, 'batched-migrations')],
+        });
+
+        runner.on('error', (err) => {
+          logger.error('Batched migration runner error', err);
+          Sentry.captureException(err);
+        });
       },
       async () => {
         // Using the `--migrate-and-exit` flag will override the value of
         // `config.runMigrations`. This allows us to use the same config when
         // running migrations as we do when we start the server.
         if (config.runMigrations || argv['migrate-and-exit']) {
-          await migrations.init(path.join(__dirname, 'migrations'), 'prairielearn');
+          await migrations.init(
+            [path.join(__dirname, 'migrations'), SCHEMA_MIGRATIONS_PATH],
+            'prairielearn'
+          );
 
           if (argv['migrate-and-exit']) {
             logger.info('option --migrate-and-exit passed, running DB setup and exiting');
@@ -2063,6 +2074,16 @@ if (config.startServer) {
         });
       },
       async () => {
+        if (config.runBatchedMigrations) {
+          // Now that all migrations have been run, we can start executing any
+          // batched migrations that may have been enqueued by migrations.
+          startBatchedMigrations({
+            workDurationMs: config.batchedMigrationsWorkDurationMs,
+            sleepDurationMs: config.batchedMigrationsSleepDurationMs,
+          });
+        }
+      },
+      async () => {
         // We create and activate a random DB schema name
         // (https://www.postgresql.org/docs/12/ddl-schemas.html)
         // after we have run the migrations but before we create
@@ -2125,12 +2146,6 @@ if (config.startServer) {
       },
       async () => await codeCaller.init(),
       async () => await assets.init(),
-      (callback) => {
-        redis.init((err) => {
-          if (ERR(err, callback)) return;
-          callback(null);
-        });
-      },
       (callback) => {
         cache.init((err) => {
           if (ERR(err, callback)) return;
@@ -2204,6 +2219,7 @@ if (config.startServer) {
             externalGraderResults.stop(),
             cron.stop(),
             serverJobs.stop(),
+            stopBatchedMigrations(),
           ]);
           results.forEach((r) => {
             if (r.status === 'rejected') {