+++ tests/testAssessmentOrdering.test.js
@@ -4,7 +4,7 @@ const assert = require('chai').assert;
 const { step } = require('mocha-steps');
 const { v4: uuid } = require('uuid');
 
-const config = require('../lib/config');
+const { config } = require('../lib/config');
 const sqldb = require('@prairielearn/postgres');
 const sql = sqldb.loadSqlEquiv(__filename);