+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/postgres
 
+## 1.6.0
+
+### Minor Changes
+
+- dbfa7a689: Allow cursor results to be consumed as a stream
+
+## 1.5.0
+
+### Minor Changes
+
+- 5d0f08ecc: Don't start nested transactions in `runInTransactionAsync`
+
+## 1.4.0
+
+### Minor Changes
+
+- ce16bede7: Add `queryCursor` and `queryValidatedCursor` functions
+
 ## 1.3.0
 
 ### Minor Changes